ravas/requirements.txt
opencv-python==4.11.0.86
PyAudio==0.2.14
librosa==0.11.0
mediapipe==0.10.21
psutil==7.1.0
PyQt6==6.9.1
pyvirtualcam==0.14.0
pyyaml==6.0.3
onnxruntime==1.23.1
selenium==4.36.0
websocket_server
moderngl-window==3.1.1
torch==2.7.1
torchaudio==2.8.0
safetensors==0.5.3
huggingface_hub==0.33.5
einops==0.8.1
onnxscript==0.5.3